Create script to add __attribute__ to some prototype functions, fix reported warnings.
Former-commit-id: 733ada7c98b34bca6a1674ae5a902d925324fe58

+++ server/src/server/login.c
@@ -420,7 +420,7 @@ void check_login(object *op)
 
 		if (elapsed_save_time < 0)
 		{
-			LOG(llevBug, "BUG: Player file %s was saved in the future? (%d time)\n", filename, elapsed_save_time);
+			LOG(llevBug, "BUG: Player file %s was saved in the future? (%lu time)\n", filename, (unsigned long) elapsed_save_time);
 			elapsed_save_time = 0;
 		}
 	}
